move function to __init__
